admin : addtext

+++ db/dbManager.go
@@ -11,7 +11,7 @@ type DbManager interface {
 
     GetIpsum(s string) (map[string]string, error)
     CreateIpsum(name string, desc string, uri string, adminEmail string) (sqlRes, error)
-    AddText(ipsumId int, text string) (sqlRes, error)
-    UpdateText(dataId int, text string) (sqlRes, error)
-    DeleteText(dataId int) (sqlRes, error)
+    AddText(ipsumId int64, text string) (sqlRes, error)
+    UpdateText(dataId int64, text string) (sqlRes, error)
+    DeleteText(dataId int64) (sqlRes, error)
 }
